requirements.txt
ChatterBot==1.0.4
imblearn==0.0
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.4
Pillow==9.0.1
plotly==5.3.1
pytz==2021.1
scikit_learn==1.0.2
seaborn==0.11.1
SQLAlchemy==1.2.19
statsmodels==0.12.2
streamlit==0.89.0
xlrd==1.2.0
openpyxl==3.0.9




